Revert "Mount sqlalchemy models to multiple services"
This reverts commit c6b79c98aa36967c36539a6bf2000aee621e15c9.
