+++ docker-compose.yml
@@ -39,7 +39,6 @@ services:
       POSTGRES_SUPER_PASSWORD: ${POSTGRES_SUPER_PASSWORD}
     volumes:
       - ./mosquitto-consumer/logs:/app/logs
-      - ./postgres/database/models.py:/app/src/mosquitto_consumer/database/models.py:ro
     depends_on:
       - postgres
       - mosquitto-broker
@@ -59,7 +58,6 @@ services:
       - postgres
     volumes:
       - ./api-server/logs:/app/logs
-      - ./postgres/database/models.py:/app/src/api_server/database/models.py:ro
     restart: unless-stopped
     ports:
       - "8000:8000"
