+++ mosquitto-consumer/src/mosquitto_consumer/config/enums.py
@@ -1,5 +1,14 @@
-from enum import Enum
+from enum import Enum, StrEnum, auto
 
+
+class TableNames(StrEnum):
+    """String enums for Postgres table names."""
+
+    _value_: auto
+
+    PLANTS = auto()
+    PLANTS_MOISTURE_LOG = auto()
+    RECOMMENDED_PLANT_MOISTURE = auto()
 
 class MosquittoSubscribeMethod(Enum):
     """String enums to determine MQTT subscription method."""
