+++ mosquitto-consumer/src/mosquitto_consumer/database/schema.sql
